requirements.txt
# Python dependencies for SaaS Analytics Demo

# Database connectivity
psycopg2-binary==2.9.9

# Optional: For enhanced data analysis and visualization
# pandas==2.1.4
# matplotlib==3.8.2
# seaborn==0.13.0
# plotly==5.17.0

# Optional: For Jupyter notebook support
# jupyter==1.0.0
# ipykernel==6.28.0

# Optional: For web dashboard development
# streamlit==1.29.0
# dash==2.15.0

# Development and testing
# pytest==7.4.3
# black==23.12.0
# flake8==6.1.0